requirements.txt
selenium
pytest
pytest-html
webdriver-manager
requests
flask
openpyxl
psutil
python-docx
Pillow
